Update 6.
Made IEnumerable_1_GetEnumerator in BaseList Protected. It should not be visible.
Added CastAsList extension method, and CastAsListCollection class.
The design of SelectAsNotifyingList was completely incorrect. It also had a bug that it got the original members in the event args of the CollectionChanged event instead of the selected members. The collection has been made lazy (no stores happen now, no calls to the select method but from the user). It is now SelectAsList, with the event.
Moved foreward the localization library. The TextBlock now supports some kind of format string. This way the localization library will be fully functional probably, if our God wills.